refactor(search): revert to trait-object CompoundSearchIndex/CompoundVectorIndex
